+++ src/zgesv_rbt_batched.cpp
@@ -140,7 +140,7 @@ magma_zgesv_rbt_batched(
     for (i=0; i < batchCount; i++)
     {
         if (cpu_info[i] != 0 ) {
-            printf("magma_zgetrf_batched matrix %ld returned error %ld\n", long(i), long(cpu_info[i]) );
+            printf("magma_zgetrf_batched matrix %lld returned error %lld\n", (long long) i, (long long) cpu_info[i] );
             info = cpu_info[i];
             magma_free_cpu (cpu_info);
             return info;
