+++ src/zheevdx_2stage_m.cpp
@@ -343,14 +343,14 @@ magma_zheevdx_2stage_m(
     }
 
 
-    timer_printf("using %ld parallel_threads\n", long(parallel_threads) );
+    timer_printf("using %lld parallel_threads\n", (long long) parallel_threads );
 
     /* Check if matrix is very small then just call LAPACK on CPU, no need for GPU */
     magma_int_t ntiles = n/nb;
     if ( ( ntiles < 2 ) || ( n <= 128 ) ) {
         #ifdef ENABLE_DEBUG
         printf("--------------------------------------------------------------\n");
-        printf("  warning matrix too small N=%ld NB=%ld, calling lapack on CPU\n", long(n), long(nb) );
+        printf("  warning matrix too small N=%lld NB=%lld, calling lapack on CPU\n", (long long) n, (long long) nb );
         printf("--------------------------------------------------------------\n");
         #endif
         lapackf77_zheevd(jobz_, uplo_, &n,
@@ -449,7 +449,7 @@ magma_zheevdx_2stage_m(
     magma_zhetrd_he2hb(uplo, n, nb, A, lda, TAU1, Wstg1, lwstg1, dT1sgpu, info);
 
     timer_stop( time );
-    timer_printf( "  N= %10ld  nb= %5ld time zhetrd_he2hb= %6.2f\n", long(n), long(nb), time );
+    timer_printf( "  N= %10lld  nb= %5lld time zhetrd_he2hb= %6.2f\n", (long long) n, (long long) nb, time );
     timer_start( time );
 #else
     magma_int_t nstream = max(3,ngpu+2);
@@ -460,7 +460,7 @@ magma_zheevdx_2stage_m(
     magma_int_t distblk = max(256, 4*nb);
 
     #ifdef ENABLE_DEBUG
-    printf("voici ngpu %ld distblk %ld NB %ld nstream %ld\n ", ngpu, distblk, nb, nstream);
+    printf("voici ngpu %lld distblk %lld NB %lld nstream %lld\n ", ngpu, distblk, nb, nstream);
     #endif
 
     timer_start( time_alloc );
@@ -527,15 +527,15 @@ magma_zheevdx_2stage_m(
     }
 
     timer_stop( time );
-    timer_printf( "  N= %10ld  nb= %5ld time zhetrd_convert = %6.2f\n", long(n), long(nb), time );
+    timer_printf( "  N= %10lld  nb= %5lld time zhetrd_convert = %6.2f\n", (long long) n, (long long) nb, time );
     timer_start( time );
 
     magma_zhetrd_hb2st(uplo, n, nb, Vblksiz, A2, lda2, W, E, V2, ldv, TAU2, wantz, T2, ldt);
 
     timer_stop( time );
     timer_stop( time_total );
-    timer_printf( "  N= %10ld  nb= %5ld time zhetrd_hb2st= %6.2f\n", long(n), long(nb), time );
-    timer_printf( "  N= %10ld  nb= %5ld time zhetrd= %6.2f\n", long(n), long(nb), time_total );
+    timer_printf( "  N= %10lld  nb= %5lld time zhetrd_hb2st= %6.2f\n", (long long) n, (long long) nb, time );
+    timer_printf( "  N= %10lld  nb= %5lld time zhetrd= %6.2f\n", (long long) n, (long long) nb, time_total );
 
     /* For eigenvalues only, call DSTERF.  For eigenvectors, first call
      ZSTEDC to generate the eigenvector matrix, WORK(INDWRK), of the
@@ -548,7 +548,7 @@ magma_zheevdx_2stage_m(
         magma_dmove_eig(range, n, W, &il, &iu, vl, vu, m);
 
         timer_stop( time );
-        timer_printf( "  N= %10ld  nb= %5ld time dstedc = %6.2f\n", long(n), long(nb), time );
+        timer_printf( "  N= %10lld  nb= %5lld time dstedc = %6.2f\n", (long long) n, (long long) nb, time );
     }
     else {
         timer_start( time_total );
@@ -572,7 +572,7 @@ magma_zheevdx_2stage_m(
 #endif
 
         timer_stop( time );
-        timer_printf( "  N= %10ld  nb= %5ld time zstedx_m = %6.2f\n", long(n), long(nb), time );
+        timer_printf( "  N= %10lld  nb= %5lld time zstedx_m = %6.2f\n", (long long) n, (long long) nb, time );
 
         magma_dmove_eig(range, n, W, &il, &iu, vl, vu, m);
 
@@ -596,7 +596,7 @@ magma_zheevdx_2stage_m(
 
 
         timer_stop( time );
-        timer_printf( "  N= %10ld  nb= %5ld time zbulge_back_m = %6.2f\n", long(n), long(nb), time );
+        timer_printf( "  N= %10lld  nb= %5lld time zbulge_back_m = %6.2f\n", (long long) n, (long long) nb, time );
         timer_start( time );
 
 #ifdef SINGLEGPU
@@ -610,14 +610,14 @@ magma_zheevdx_2stage_m(
 #endif // not SINGLEGPU
 
         timer_stop( time );
-        timer_printf( "  N= %10ld  nb= %5ld time zunmqr_m + copy = %6.2f\n", long(n), long(nb), time );
+        timer_printf( "  N= %10lld  nb= %5lld time zunmqr_m + copy = %6.2f\n", (long long) n, (long long) nb, time );
 #ifdef SINGLEGPU
         magma_free(dT1sgpu);
         magma_free(dZ);
 #endif // not SINGLEGPU
 
         timer_stop( time_total );
-        timer_printf( "  N= %10ld  nb= %5ld time eigenvectors backtransf. = %6.2f\n", long(n), long(nb), time_total );
+        timer_printf( "  N= %10lld  nb= %5lld time eigenvectors backtransf. = %6.2f\n", (long long) n, (long long) nb, time_total );
     }
 
     /* If matrix was scaled, then rescale eigenvalues appropriately. */
